requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langchain-community>=0.0.10

# AI & LLM
openai>=1.10.0
anthropic>=0.8.0

# Search & Data
tavily-python>=0.3.0

# Data & Validation
pydantic>=2.5.0

# Utilities
python-dotenv>=1.0.0

# Vector DB (optional, for future RAG implementation)
chromadb>=0.4.0

# Web scraping (optional)
beautifulsoup4>=4.12.0
requests>=2.31.0
